Refactored css into separate file
But to do so I had to serve static assets using chi-router so that the
css files could be found. (That adds more complexity, but is unavoidable
since I would have to serve js files the same way as well.)

Also added the skeleton if/else template logic for switching the navbar
between public/participant/adviser/mentor/admin mode depending on what
variables you pass to the template.

TODO: properly flesh out the navbar for each mode (right now they are
all identical). Also figure out the correct data struct to pass to the
navbar template. The struct will provide some sort of guideline for what
variables to pass to the templates, so it's not like you have to trace
the entire template paths in order to figure it out at least.