Laid out the groundwork on how to organize and pass struct data into the templates
Also wired up the navbar display logic as mentioned in the previous
commit.

Made a change of heart when I realized I wanted the navbar html file to
be contain all the if/else display logic instead of embedding it in the
calling html file. So now it's called 'navbar.html' instead of
'navbars.html' because it contains THE navbar, which changes depending
on what variables you pass to it (as compared to simply being a file
that contains all the templates of the various navbar states).

Reindented css file to 2 spaces. Added static/orbital.jpg.