Automate populating db with test data using data.sql
Note: insert_submission stored procedure is needed so that the MD5 hash
of CONCAT(team, milestone, project_level, project_link, project_readme,
project_poster, project_video) is automatically calculated and inserted
together with the rest of the row. This is very cool, technically you
don't even need to implement the MD5 hashing in the go code anymore
because you can just call the stored procedure and the database will do
it for you.