uid foreign key in tutors/advisers/..etc is no longer unique
Instead, it is the combination of uid and cohort that is unique. This
allows the same user to serve as a tutor or adviser across multiple
cohorts.

This also means that 'adviser' columns in tables like
adviser_evaluationships can no longer reference an adviser table uid (no
uniqueness guarantee), but rather a users uid. Which means that someone
could manually insert a non-adviser user into that table. This would
have to be dealt with using a further existence constraint, but I'm too
lazy to figure it out right now. As long as no one manually tries to
sabotage the table it should be fine, but keep in view.

Also remembered about database views (thanks CS2102!) which act like
stored SELECT statements. Views would allow me to move SQL logic from
the application itself into the database, which is always nice.

However, the views are part of a larger problem here. The way I have
denormalized the table, it makes it a pain in the ass to get the
following details in a single JOINed table:
• Project name/link/poster/video
• teammate 1
• teammate 2
• adviser
• mentor

There are two reasons why

1) I can look at a team member and tell which team he's in, but I cannot
look at a team and tell who the team members are. This is because the
team-to-teammember info is stored in the participants table, not the
teams table.

2) I can't use the participant/adviser/...etc tables at all without
doing a JOIN with the users table, because all the user details are in
the users table and not the participant/adviser/...etc tables. Which
makes for an annoying overhead for everything.

I have solved those problems by using views, but there's just one
problem: It could be slow. It could be like doing an SQL equivalent of
O(N^3). Right now the vvv_teams view contains the required details I
mentioned above, but it contains three levels of views, and the internet
says that nested views are bad. But I can't find any other way around
this.

The solution is to decouple the database implementation and details so
cleanly from the application that it's easy to refactor in the future.
The entire database's table schemas might change. But for now I'll keep
using this perhaps-slow SQL query.