MVP of a Model-View-Controller architecture
Just committing this in because I'm thinking of switching to a leaner
Frontend-Contract-Backend architecture: I don't really see the point of
a Controller when the Model can do everything.

The idea behind MVC is that the contract between the View and the
Controller is different from the contract between the Controller and the
Model: So the Controller serves to translate data between the two
contracts.

But the contract between the Model and Controller is almost identical
(or exactly identical) to the contract between the Controller and the
View. So I plan to meld the Controller and Model into one, called the
Backend, with the View becoming the Frontend.
For each template, the Frontend and the Backend will agree on a list of
variables to read from/ deliver to, codified as a Contract (essentially
a struct of variables).

How would this make the database easier to radically refactor? Well if
the database changes or some tables change schemas entirely, all the
Backend has to do is to continue delivering that list of variables
specified in the Contract, by hook or by crook. It makes it clear where
its endpoint should be.

However the presence of the user authentication system could throw a
spanner in the works, so I'm going to investigate on how to create a
user account system first before fully committing to this
Frontend-Contract-Backend architecture.