Every page now has a hash identifier tied to it
The hash guarantees that any variable/function/file name that contains
that hash will be related to that page.  Instead of thinking of a
meaningful semantic name that won't clash, then appending that (possibly
long) name to the end of relevant functions and variables, you can just
append its hash. Grepping through the codebase for that hash would show
you the relevant variables easily.

The hash is the first 7 digits of a SHA1 hash of a string, usually of
the semantic name itself. For example, the Past Year Showcase page has
the hash 361d489 which is generated by

echo "pastyearshowcase" | sha1sum | awk '{printf "%s",$1}' | head -c 7

Generating the hash off a semantic name isn't a requirement, it just
reduces the chance of a hash collision comapred to hashing it with
common names. You could also generate a hash with a long and nonsensical
string.

Now I can add the hash inside every variable/function/file name that is
related to the Past Year Showcase page. There are a few guarantees for a
<hash>-identified page:
    The controller function is called *<hash>(), and it belongs inside a *.<hash>.go file
    The Contract between the controller function and the parent template is stored inside a Contract<hash>{} struct
    The parent template is a *.<hash>.html file
    The styling belongs to a *.<hash>.css file
    The interactivity is served by a *.<hash>.js file
To illustrate, the Past Year Showcase page
    is routed by PastYearShowcase361d489()
    which is located inside pastYearShowcase.361d489.go
    which fills up a struct called Contract361d489
    and then sends it to the template pastYearShowcase.361d489.html for rendering
    (styling is inside pastYearShowcase.361d489.css)

Again, this is just to make the dataflow for a particular page explicit.
This came about as a concern because some of the pages I plan to
implement are called 'Register' and 'Login', which are very ripe for
function names as well. I wanted a way to differentiate any possible
function name called 'Register' and the 'Register' title that would
identify a page's .go, .html & .css files. So my solution was to add a
unique id behind every title. Also has the benefit of being able to use
the hash as a short form in place of the longer semantic name (e.g.
Contract361d489 instead of ContractPastYearShowcase)